GTM19-565: Fix evaluation of EXTW
GNUSim evaluates EXTW incorrectly when the first input register is
also used as the output and that value in that register has MSB set.
Due to sign extension, the upper 32-bits in the 64-bit (internal)
representation of that register is all 1s.  These upper bits swamp
out the bits coming from the 2nd register.

sim/
	* mips/nanomipsr6.igen <ALIGN|EXTW>: Remove incorrect special
	handling of EXTW when the first source register is also the
	destination.