CR MOLY00690538: Reloc-minimization introduces incorrect branch target
The relax_sink field in tc_frag_data tracks the minimum possible offset
of a branch instruction subjected to relaxation.  This field should
accumulate all sinks between the branch source and target.  When
referring to an external symbol, relax_sink is incorrectly being
assigned instead of being added to.

This leads to branches which are link-time  variable appearing
invariable and the target encoded by the assembler turns out to be
incorrect after linker relaxation.

gas/
	* config/tc-nanomips.c (relaxed_invariable_branch_p): Accumulate
	variant frag size in to relax_sink for external references.
	* testsuite/gas/nanomips/minimize-relocs6.s: New source.
	* testsuite/gas/nanomips/minimize-relocs6.d: New test.
	* testsuite/gas/nanomips/nanomips.exp: Run the new test.

Ref: http://172.21.160.85/swits/mtkRecord.wits?method=redirectPage&recordType=CR&action=View&crId=MOLY00690538